refactor(core): consolidate tmux access into a single core/tmux.ts adapter
tmux was shelled out from five places (runtime driver, runs probes, cli
TUI-launcher helpers, doctor version check, web bridge), each re-deriving
the =-exact-match rule and the (session name, id) identity doctrine. The
stale-id fix (#17) had to touch all five, which is the cost this removes.

core/tmux.ts now owns the RuntimeDriver seam, realDriver, the identity
doctrine comment, and every generic probe: tmuxAvailable, tmuxVersion,
sessionExists, targetExists, windowIds, windowByName, windowInSession,
paneInSession, exactSession, STALE_WINDOW_ERROR. Consumers import from
it; runtime.ts re-exports the RuntimeDriver type so its importers are
unchanged. TUI-session policy helpers stay in cli.ts. No behavior
change: same argv, same injectable seams, test counts identical.
